Remove leftover cruft: delete unused LazyFrameBatchIter, revert uv.lock churn
- Delete xgboost_forecaster/_data_iter.py (LazyFrameBatchIter, an abandoned
  row-batching DataIter approach) and its tests; move the still-used
  _prepare_features helper into forecaster.py. Drop the doc/README references to
  the iterator.
- Fix a stale BaseForecaster docstring ("one spatial cell at a time" → "one
  init_time chunk at a time") left over from the per-cell prediction attempt.
- Revert uv.lock: pyproject deps are unchanged in this PR, so the transitive
  version churn (polars 1.42.0→1.42.1, etc.) doesn't belong here.

train(data, time_series_ids) is kept deliberately: the population is used to
restrict which series are fitted and is the train==predict invariant (and the
hook for future one-booster-per-group models).
